refactor(UserAPIController): Extract RunInfoController
