+++ src/main/java/org/wise/portal/presentation/web/controllers/user/UserAPIController.java
@@ -230,49 +230,6 @@ public class UserAPIController {
     return langs;
   }
 
-  @Secured("ROLE_USER")
-  @GetMapping("/run/info-by-id")
-  HashMap<String, Object> getRunInfoById(Authentication auth, @RequestParam("runId") Long runId) {
-    try {
-      User user = userService.retrieveUserByUsername(auth.getName());
-      Run run = runService.retrieveById(runId);
-      if (userService.isUserAssociatedWithRun(user, run)) {
-        return getRunInfo(run);
-      }
-    } catch (ObjectNotFoundException e) {
-    }
-    return createRunNotFoundInfo();
-  }
-
-  protected HashMap<String, Object> getRunInfo(Run run) {
-    HashMap<String, Object> info = new HashMap<String, Object>();
-    info.put("id", String.valueOf(run.getId()));
-    info.put("name", run.getName());
-    info.put("runCode", run.getRuncode());
-    info.put("startTime", run.getStartTimeMilliseconds());
-    info.put("endTime", run.getEndTimeMilliseconds());
-    info.put("periods", getPeriodNames(run));
-    User owner = run.getOwner();
-    info.put("teacherFirstName", owner.getUserDetails().getFirstname());
-    info.put("teacherLastName", owner.getUserDetails().getLastname());
-    info.put("wiseVersion", run.getProject().getWiseVersion());
-    return info;
-  }
-
-  private List<String> getPeriodNames(Run run) {
-    List<String> periods = new ArrayList<String>();
-    for (Group period : run.getPeriods()) {
-      periods.add(period.getName());
-    }
-    return periods;
-  }
-
-  protected HashMap<String, Object> createRunNotFoundInfo() {
-    HashMap<String, Object> info = new HashMap<String, Object>();
-    info.put("error", "runNotFound");
-    return info;
-  }
-
   private String getLanguageName(String localeString) {
     if (localeString.toLowerCase().equals("zh_tw")) {
       return "Chinese (Traditional)";
