+++ src/test/java/org/wise/portal/presentation/web/controllers/student/StudentAPIControllerTest.java
@@ -14,13 +14,11 @@ import java.util.Properties;
 import java.util.Set;
 
 import org.easymock.EasyMockExtension;
-import org.easymock.EasyMockRunner;
 import org.easymock.Mock;
 import org.easymock.TestSubject;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.runner.RunWith;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.util.ReflectionTestUtils;
@@ -122,29 +120,7 @@ public class StudentAPIControllerTest extends APIControllerTest {
     verify(request);
   }
 
-  @Test
-  public void getRunInfoByRunCode_RunExistsInDB_ReturnRunInfo() throws ObjectNotFoundException {
-    expect(runService.retrieveRunByRuncode(RUN1_RUNCODE)).andReturn(run1);
-    replay(runService);
-    HashMap<String, Object> info = studentAPIController.getRunInfoByRunCode(RUN1_RUNCODE);
-    assertEquals("1", info.get("id"));
-    assertEquals(RUN1_RUNCODE, info.get("runCode"));
-    verify(runService);
-  }
-
-  @Test
-  public void getRunInfoByRunCode_RunNotInDB_ReturnRunInfo() throws ObjectNotFoundException {
-    String runCodeNotInDB = "runCodeNotInDB";
-    expect(runService.retrieveRunByRuncode(runCodeNotInDB))
-        .andThrow(new ObjectNotFoundException(runCodeNotInDB, Run.class));
-    replay(runService);
-    HashMap<String, Object> info = studentAPIController.getRunInfoByRunCode(runCodeNotInDB);
-    assertEquals(1, info.size());
-    assertEquals("runNotFound", info.get("error"));
-    verify(runService);
-  }
-
-  @Test
+ @Test
   public void getSecurityQuestions_DefaultQuestions_ReturnSixQuestions() {
     expect(i18nProperties.getProperty("accountquestions.QUESTION_ONE"))
         .andReturn("account question 1");
